qt5: build under c++11 mode
if we are building against qt5, build under c++11 mode for binary consistence.
this requires your compiler newer than gcc 4.8.1 or clang 3.3.
and, currently, it is guarenteed under osx.

from qt5.4.0's changelog https://qt.gitorious.org/qt/qtbase/source/64c6bfd1f45f532ea32de49431b66e03e7593f01:dist/changes-5.4.0#L443-449
"The Qt binary packages (for OS X) are now configured with C++11 enabled."

+++ src/filebrowser/file-browser-dialog.cpp
@@ -650,7 +650,7 @@ void FileBrowserDialog::onUploadFinished(bool success)
           SeafDirent::FILE,
           "",
           name,
-          file.size(),
+          static_cast<quint64>(file.size()),
           QDateTime::currentDateTime().toTime_t()
         };
         if (task->useUpload())
